+++ Nemesis-Mod/src/bluetooth/ConfigurationService.cpp
