+++ Nemesis-Mod/src/bluetooth/CustomBLEService.cpp
